pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "everybody-codes"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "tobeannouncd", email = "wzrds3@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "requests",
  "pycryptodome",
]

[project.urls]
Documentation = "https://github.com/tobeannouncd/everybody-codes-api#readme"
Issues = "https://github.com/tobeannouncd/everybody-codes-api/issues"
Source = "https://github.com/tobeannouncd/everybody-codes-api"

[tool.hatch.version]
path = "src/everybody_codes/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/everybody_codes tests}"

[tool.coverage.run]
source_pkgs = ["everybody_codes", "tests"]
branch = true
parallel = true
omit = [
  "src/everybody_codes/__about__.py",
]

[tool.coverage.paths]
everybody_codes = ["src/everybody_codes", "*/everybody-codes/src/everybody_codes"]
tests = ["tests", "*/everybody-codes/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
